TSDK-790 Refactor for testing
