+++ shared/src/main/scala/co/topl/shared/ParamParser.scala
@@ -1,5 +1,12 @@
 package co.topl.shared
 
+import org.bitcoins.core.currency.CurrencyUnit
+
+import org.bitcoins.core.currency.SatoshisLong
+import scala.util.Try
+import scala.util.Success
+import scala.util.Failure
+
 object ParamParser {
 
   implicit val networkRead: scopt.Read[BitcoinNetworkIdentifiers] =
@@ -23,6 +30,15 @@ object ParamParser {
             "Invalid network. Possible values: mainnet, testnet, private"
           )
       })
-
+  implicit val currencyUnit: scopt.Read[CurrencyUnit] =
+    scopt.Read
+      .reads(x => Try(x.toLong.satoshis))
+      .map(_ match {
+        case Success(v) => v
+        case Failure(_) =>
+          throw new IllegalArgumentException(
+            "Could not conver value to satoshi"
+          )
+      })
 
 }
