+++ topl-btc-bridge/src/main/scala/co/topl/bridge/managers/AssetMintingOps.scala
@@ -1,7 +1,5 @@
 package co.topl.bridge.managers
 
-
-
 import cats.effect.kernel.Sync
 import co.topl.brambl.builders.TransactionBuilderApi
 import co.topl.brambl.dataApi.WalletStateAlgebra
@@ -19,6 +17,7 @@ import io.circe.Json
 import quivr.models.KeyPair
 
 import TransactionBuilderApi.implicits._
+import co.topl.bridge.Lvl
 
 trait AssetMintingOps[G[_]] extends CommonTxOps {
 
@@ -40,7 +39,7 @@ trait AssetMintingOps[G[_]] extends CommonTxOps {
       seriesTxo: Txo,
       lockAddrToUnlock: LockAddress,
       lockPredicateFrom: Lock.Predicate,
-      fee: Long,
+      fee: Lvl,
       someNextIndices: Option[Indices],
       assetMintingStatement: AssetMintingStatement,
       ephemeralMetadata: Option[Json],
@@ -81,12 +80,13 @@ trait AssetMintingOps[G[_]] extends CommonTxOps {
       lockPredicateFrom: Map[LockAddress, Lock.Predicate],
       lockForChange: Lock,
       recipientLockAddress: LockAddress,
-      fee: Long,
+      fee: Lvl,
       assetMintingStatement: AssetMintingStatement,
       ephemeralMetadata: Option[Struct],
       commitment: Option[ByteString],
       someNextIndices: Option[Indices]
-  ): G[IoTransaction] =
+  ): G[IoTransaction] = {
+    import co.topl.brambl.syntax._
     for {
       changeAddress <- tba.lockAddress(
         lockForChange
@@ -95,7 +95,7 @@ trait AssetMintingOps[G[_]] extends CommonTxOps {
         assetMintingStatement,
         txos,
         lockPredicateFrom,
-        fee,
+        fee.amount.toLong,
         recipientLockAddress,
         changeAddress,
         ephemeralMetadata,
@@ -126,4 +126,5 @@ trait AssetMintingOps[G[_]] extends CommonTxOps {
           Sync[G].delay(())
         }
     } yield ioTransaction
+  }
 }
