+++ topl-btc-bridge/src/main/scala/co/topl/bridge/modules/InitializationModule.scala
@@ -20,6 +20,8 @@ import quivr.models.KeyPair
 
 import scala.concurrent.duration._
 import co.topl.bridge.SystemGlobalState
+import co.topl.bridge.Fellowship
+import co.topl.bridge.Template
 
 class InitializationModule[F[_]: Async: Logger](
     val tba: TransactionBuilderApi[F],
@@ -41,10 +43,10 @@ class InitializationModule[F[_]: Async: Logger](
 
   val m: Monad[F] = implicitly[Monad[F]]
 
-  val fromFellowship = "self"
-  val fromTemplate = "default"
-
-  private def getTxos(): F[Seq[Txo]] = for {
+  private def getTxos(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Seq[Txo]] = for {
     currentAddress <- getCurrentAddress(
       fromFellowship,
       fromTemplate,
@@ -65,10 +67,13 @@ class InitializationModule[F[_]: Async: Logger](
       .fold(longAsInt128(0))(_ + _)
   }
 
-  def checkForLvls(): F[Unit] = (for {
+  def checkForLvls(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Unit] = (for {
     _ <- info"Checking for LVLs"
     currentAddress <- wsa.getCurrentAddress
-    txos <- getTxos()
+    txos <- getTxos(fromFellowship, fromTemplate)
     hasLvls <-
       if (txos.filter(_.transactionOutput.value.value.isLvl).nonEmpty) {
         (info"Found LVLs: ${int128AsBigInt(sumLvls(txos))}" >> currentState
@@ -95,7 +100,7 @@ class InitializationModule[F[_]: Async: Logger](
       }
     _ <-
       if (!hasLvls)
-        Async[F].sleep(5.second) >> checkForLvls()
+        Async[F].sleep(5.second) >> checkForLvls(fromFellowship, fromTemplate)
       else Async[F].unit
   } yield ()).handleErrorWith { e =>
     e.printStackTrace()
@@ -103,11 +108,14 @@ class InitializationModule[F[_]: Async: Logger](
       error"Retrying in 5 seconds" >>
       Async[F].sleep(
         5.second
-      ) >> checkForLvls()
+      ) >> checkForLvls(fromFellowship, fromTemplate)
   }
 
-  private def checkIfGroupTokenMinted(): F[Unit] = for {
-    newTxos <- getTxos()
+  private def checkIfGroupTokenMinted(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Unit] = for {
+    newTxos <- getTxos(fromFellowship, fromTemplate)
     _ <-
       if (newTxos.filter(_.transactionOutput.value.value.isGroup).nonEmpty) {
         info"Group Token minted successfully" >> currentState
@@ -126,12 +134,18 @@ class InitializationModule[F[_]: Async: Logger](
               currentError = None,
               isReady = false
             )
-          ) >> Async[F].sleep(5.second) >> checkIfGroupTokenMinted()
+          ) >> Async[F].sleep(5.second) >> checkIfGroupTokenMinted(
+          fromFellowship,
+          fromTemplate
+        )
       }
   } yield ()
 
-  def checkIfSeriesTokenMinted(): F[Unit] = for {
-    newTxos <- getTxos()
+  def checkIfSeriesTokenMinted(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Unit] = for {
+    newTxos <- getTxos(fromFellowship, fromTemplate)
     _ <-
       if (newTxos.filter(_.transactionOutput.value.value.isSeries).nonEmpty) {
         info"Series Token minted successfully" >> currentState
@@ -150,13 +164,19 @@ class InitializationModule[F[_]: Async: Logger](
               currentError = None,
               isReady = false
             )
-          ) >> Async[F].sleep(5.second) >> checkIfSeriesTokenMinted()
+          ) >> Async[F].sleep(5.second) >> checkIfSeriesTokenMinted(
+          fromFellowship,
+          fromTemplate
+        )
       }
   } yield ()
 
-  def mintGroupToken(): F[Unit] = (for {
+  def mintGroupToken(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Unit] = (for {
     _ <- info"Minting Group Token"
-    txos <- getTxos()
+    txos <- getTxos(fromFellowship, fromTemplate)
     lockAddress <- wsa.getCurrentAddress
     someLock <- wsa.getLockByAddress(lockAddress)
     _ = assert(
@@ -168,8 +188,8 @@ class InitializationModule[F[_]: Async: Logger](
       "No LVLs found while minting group token"
     )
     someChangeIdx <- wsa.getNextIndicesForFunds(
-      fromFellowship,
-      fromTemplate
+      fromFellowship.underlying,
+      fromTemplate.underlying
     )
     _ = assert(
       someChangeIdx.isDefined,
@@ -217,7 +237,7 @@ class InitializationModule[F[_]: Async: Logger](
       provedTx.toOption.get
     )
     _ <- Async[F].fromEither(eitherSentTx) // if this fails we should retry
-    _ <- checkIfGroupTokenMinted()
+    _ <- checkIfGroupTokenMinted(fromFellowship, fromTemplate)
 
   } yield ()).handleErrorWith { e =>
     e.printStackTrace()
@@ -225,12 +245,15 @@ class InitializationModule[F[_]: Async: Logger](
       error"Retrying in 5 seconds" >>
       Async[F].sleep(
         5.second
-      ) >> mintGroupToken()
+      ) >> mintGroupToken(fromFellowship, fromTemplate)
   }
 
-  def mintSeriesToken(): F[Unit] = (for {
+  def mintSeriesToken(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Unit] = (for {
     _ <- info"Minting Series Token"
-    txos <- getTxos()
+    txos <- getTxos(fromFellowship, fromTemplate)
     lockAddress <- wsa.getCurrentAddress
     someLock <- wsa.getLockByAddress(lockAddress)
     _ = assert(
@@ -242,8 +265,8 @@ class InitializationModule[F[_]: Async: Logger](
       "No Group Tokens found while minting series token"
     )
     someChangeIdx <- wsa.getNextIndicesForFunds(
-      fromFellowship,
-      fromTemplate
+      fromFellowship.underlying,
+      fromTemplate.underlying
     )
     _ = assert(
       someChangeIdx.isDefined,
@@ -291,7 +314,7 @@ class InitializationModule[F[_]: Async: Logger](
       provedTx.toOption.get
     )
     _ <- Async[F].fromEither(eitherSentTx) // if this fails we should retry
-    _ <- checkIfSeriesTokenMinted()
+    _ <- checkIfSeriesTokenMinted(fromFellowship, fromTemplate)
 
   } yield ()).handleErrorWith { e =>
     e.printStackTrace
@@ -299,13 +322,16 @@ class InitializationModule[F[_]: Async: Logger](
       error"Retrying in 5 seconds" >>
       Async[F].sleep(
         5.second
-      ) >> mintSeriesToken()
+      ) >> mintSeriesToken(fromFellowship, fromTemplate)
   }
 
-  def checkForGroupToken(): F[Boolean] = (
+  def checkForGroupToken(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Boolean] = (
     for {
       _ <- info"Checking for Group Tokens"
-      txos <- getTxos()
+      txos <- getTxos(fromFellowship, fromTemplate)
       groupTxos = txos.filter(_.transactionOutput.value.value.isGroup)
       hasGroupToken <-
         if (groupTxos.nonEmpty) {
@@ -334,10 +360,13 @@ class InitializationModule[F[_]: Async: Logger](
     } yield hasGroupToken
   )
 
-  def checkForSeriesToken(): F[Boolean] = (
+  def checkForSeriesToken(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Boolean] = (
     for {
       _ <- info"Checking for Series Tokens"
-      txos <- getTxos()
+      txos <- getTxos(fromFellowship, fromTemplate)
       hasSeriesToken <-
         if (txos.filter(_.transactionOutput.value.value.isSeries).nonEmpty) {
           (info"Found Series Tokens: ${int128AsBigInt(sumLvls(txos))}" >> currentState
@@ -363,20 +392,27 @@ class InitializationModule[F[_]: Async: Logger](
     } yield hasSeriesToken
   )
 
-  def setupWallet(): F[Unit] = {
+  def setupWallet(
+      fromFellowship: Fellowship,
+      fromTemplate: Template
+  ): F[Unit] = {
     (for {
-      _ <- checkForLvls()
-      hasGroupToken <- checkForGroupToken()
-      _ <- if (!hasGroupToken) mintGroupToken() else Async[F].unit
-      hasSeriesToken <- checkForSeriesToken()
-      _ <- if (!hasSeriesToken) mintSeriesToken() else Async[F].unit
+      _ <- checkForLvls(fromFellowship, fromTemplate)
+      hasGroupToken <- checkForGroupToken(fromFellowship, fromTemplate)
+      _ <-
+        if (!hasGroupToken) mintGroupToken(fromFellowship, fromTemplate)
+        else Async[F].unit
+      hasSeriesToken <- checkForSeriesToken(fromFellowship, fromTemplate)
+      _ <-
+        if (!hasSeriesToken) mintSeriesToken(fromFellowship, fromTemplate)
+        else Async[F].unit
     } yield ()).handleErrorWith { e =>
       e.printStackTrace
       error"Error setting up wallet: $e" >>
         error"Retrying in 5 seconds" >>
         Async[F].sleep(
           5.second
-        ) >> setupWallet()
+        ) >> setupWallet(fromFellowship, fromTemplate)
     }
   }
